+++ src/meta/measurement/systemOfMeasurement/index.ts
@@ -1,2 +1,3 @@
 export type { SystemOfMeasurement } from './systemOfMeasurement'
 export { SystemOfMeasurementName } from './systemOfMeasurementName'
+export { systemsOfMeasurement } from './systemsOfMeasurement'
